+++ backend/core/src/listings/types/listing-filter-keys-enum.ts
@@ -3,4 +3,5 @@ export enum ListingFilterKeys {
   status = "status",
   name = "name",
   neighborhood = "neighborhood",
+  bedrooms = "bedrooms",
 }